refactor(db): split monolithic db.py into a structured package
Convert strides_ai/db.py (596 lines) into a proper db/ package:

- models.py — SQLModel table definitions
- engine.py — engine singleton, init_db(), get_session() dependency
- activities/conversations/memories/settings/profiles/calendar.py — domain CRUD
- __init__.py — backward-compat wrappers for CLI/coach/sync callers

All FastAPI routers now use `session: Session = Depends(get_session)` for
request-scoped session injection. save_analysis() raw SQL replaced with
ORM setattr/session.add(). chat.py background thread creates its own
Session explicitly to avoid cross-thread session sharing.

Also update tests/conftest.py to patch db.engine.DB_PATH in addition to
db.DB_PATH so the tmp_db fixture correctly redirects the engine singleton.
