+++ strides_ai/api/routers/memories.py
@@ -1,12 +1,14 @@
 """Memories route."""
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
+from sqlmodel import Session
 
-from ... import db
+from ...db import memories as crud
+from ...db.engine import get_session
 
 router = APIRouter()
 
 
 @router.get("/memories")
-def memories():
-    return db.get_all_memories()
+def get_memories(session: Session = Depends(get_session)):
+    return [r.model_dump() for r in crud.get_all(session)]
