+++ strides_ai/api/routers/profile.py
@@ -1,9 +1,11 @@
 """Profile routes."""
 
-from fastapi import APIRouter, Request
+from fastapi import APIRouter, Depends, Request
 from pydantic import BaseModel
+from sqlmodel import Session
 
-from ... import db
+from ...db import profiles as crud
+from ...db.engine import get_session
 from ...profile import get_default_fields
 from ..deps import init_backend
 
@@ -15,24 +17,37 @@ class ProfileBody(BaseModel):
 
 
 @router.get("/profile")
-def get_profile(request: Request, mode: str | None = None):
+def get_profile(
+    request: Request,
+    mode: str | None = None,
+    session: Session = Depends(get_session),
+):
     m = mode or request.app.state.mode
-    fields = db.get_profile_fields(m) or get_default_fields(m)
+    fields = crud.get_fields(session, m) or get_default_fields(m)
     return {"fields": fields}
 
 
 @router.put("/profile")
-def put_profile(request: Request, body: ProfileBody, mode: str | None = None):
+def put_profile(
+    request: Request,
+    body: ProfileBody,
+    mode: str | None = None,
+    session: Session = Depends(get_session),
+):
     m = mode or request.app.state.mode
-    db.save_profile_fields(m, body.fields)
+    crud.save_fields(session, m, body.fields)
     init_backend(request.app)
     return {"status": "ok"}
 
 
 @router.post("/profile/reset")
-def reset_profile(request: Request, mode: str | None = None):
+def reset_profile(
+    request: Request,
+    mode: str | None = None,
+    session: Session = Depends(get_session),
+):
     m = mode or request.app.state.mode
     fields = get_default_fields(m)
-    db.save_profile_fields(m, fields)
+    crud.save_fields(session, m, fields)
     init_backend(request.app)
     return {"fields": fields}
