+++ strides_ai/api/routers/settings.py
@@ -1,10 +1,12 @@
 """Settings and provider routes."""
 
-from fastapi import APIRouter, HTTPException, Request
+from fastapi import APIRouter, Depends, HTTPException, Request
 from pydantic import BaseModel
+from sqlmodel import Session
 
-from ... import db
 from ...config import VALID_MODES, VALID_PROVIDERS, get_settings
+from ...db import settings as crud
+from ...db.engine import get_session
 from ..deps import _get_provider_models, _provider_statuses, init_backend
 
 router = APIRouter()
@@ -17,36 +19,40 @@ class SettingsBody(BaseModel):
 
 
 @router.get("/settings")
-def get_api_settings():
+def get_api_settings(session: Session = Depends(get_session)):
     return {
-        "mode": db.get_setting("mode", "running"),
-        "provider": db.get_setting("provider", get_settings().provider),
+        "mode": crud.get(session, "mode", "running"),
+        "provider": crud.get(session, "provider", get_settings().provider),
     }
 
 
 @router.put("/settings")
-def put_settings(request: Request, body: SettingsBody):
+def put_settings(
+    request: Request,
+    body: SettingsBody,
+    session: Session = Depends(get_session),
+):
     settings = get_settings()
     if body.mode is not None:
         if body.mode not in VALID_MODES:
             raise HTTPException(
                 status_code=400, detail=f"mode must be one of {sorted(VALID_MODES)}"
             )
-        db.set_setting("mode", body.mode)
+        crud.set(session, "mode", body.mode)
     if body.provider is not None:
         if body.provider not in VALID_PROVIDERS:
             raise HTTPException(
                 status_code=400, detail=f"provider must be one of {sorted(VALID_PROVIDERS)}"
             )
-        db.set_setting("provider", body.provider)
+        crud.set(session, "provider", body.provider)
     if body.model is not None:
-        target = body.provider or db.get_setting("provider", settings.provider) or "claude"
-        db.set_setting(f"{target}_model", body.model)
+        target = body.provider or crud.get(session, "provider", settings.provider) or "claude"
+        crud.set(session, f"{target}_model", body.model)
     if body.mode is not None or body.provider is not None or body.model is not None:
         init_backend(request.app, mode=body.mode, provider=body.provider)
     return {
-        "mode": db.get_setting("mode", "running"),
-        "provider": db.get_setting("provider", settings.provider),
+        "mode": crud.get(session, "mode", "running"),
+        "provider": crud.get(session, "provider", settings.provider),
     }
